+++ tests/TestCase/Controller/Admin/BannersControllerTest.php
@@ -23,10 +23,12 @@
 namespace MeCms\Test\TestCase\Controller\Admin;
 
 use Cake\Cache\Cache;
+use Cake\Controller\ComponentRegistry;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\IntegrationTestCase;
 use MeCms\Controller\Admin\BannersController;
 use MeCms\TestSuite\Traits\AuthMethodsTrait;
+use MeTools\Controller\Component\UploaderComponent;
 
 /**
  * BannersControllerTest class
@@ -46,11 +48,6 @@ class BannersControllerTest extends IntegrationTestCase
     protected $Controller;
 
     /**
-     * @var array
-     */
-    protected $url;
-
-    /**
      * Fixtures
      * @var array
      */
@@ -58,6 +55,11 @@ class BannersControllerTest extends IntegrationTestCase
         'plugin.me_cms.banners',
         'plugin.me_cms.banners_positions',
     ];
+
+    /**
+     * @var array
+     */
+    protected $url;
 
     /**
      * Setup the test case, backup the static object values so they can be
@@ -89,6 +91,34 @@ class BannersControllerTest extends IntegrationTestCase
         parent::tearDown();
 
         unset($this->Banners, $this->Controller);
+    }
+
+    /**
+     * Adds additional event spies to the controller/view event manager
+     * @param \Cake\Event\Event $event A dispatcher event
+     * @param \Cake\Controller\Controller|null $controller Controller instance
+     * @return void
+     */
+    public function controllerSpy($event, $controller = null)
+    {
+        //Sets key for cookies
+        $controller->Cookie->config('key', 'somerandomhaskeysomerandomhaskey');
+
+        //Mocks the `Uploader` component
+        $controller->Uploader = $this->getMockBuilder(UploaderComponent::class)
+            ->setConstructorArgs([new ComponentRegistry])
+            ->getMock();
+
+        $controller->Uploader->method('set')
+            ->will($this->returnSelf());
+
+        $controller->Uploader->method('mimetype')
+            ->will($this->returnSelf());
+
+        $controller->Uploader->method('save')
+            ->will($this->returnValue('/full/path/to/file_to_upload.jpg'));
+
+        parent::controllerSpy($event, $controller);
     }
 
     /**
@@ -140,15 +170,148 @@ class BannersControllerTest extends IntegrationTestCase
     }
 
     /**
+     * Tests for `index()` method
+     * @test
+     */
+    public function testIndex()
+    {
+        $this->get(array_merge($this->url, ['action' => 'index']));
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Banners/index.ctp');
+
+        $bannersFromView = $this->viewVariable('banners');
+        $this->assertInstanceof('Cake\ORM\ResultSet', $bannersFromView);
+        $this->assertNotEmpty($bannersFromView);
+
+        foreach ($bannersFromView as $banner) {
+            $this->assertInstanceof('MeCms\Model\Entity\Banner', $banner);
+        }
+
+        $this->assertCookie(null, 'renderBanners');
+    }
+
+    /**
+     * Tests for `index()` method, render as `grid`
+     * @test
+     */
+    public function testIndexAsGrid()
+    {
+        $this->get(array_merge($this->url, ['action' => 'index', '?' => ['render' => 'grid']]));
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Banners/index_as_grid.ctp');
+
+        $this->assertCookie('grid', 'renderBanners');
+    }
+
+    /**
+     * Tests for `index()` method, render as `grid` with cookie
+     * @test
+     */
+    public function testIndexWithCookie()
+    {
+        $this->cookie('renderBanners', 'grid');
+
+        $this->get(array_merge($this->url, ['action' => 'index']));
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Banners/index_as_grid.ctp');
+
+        $this->assertCookie('grid', 'renderBanners');
+    }
+
+    /**
+     * Tests for `upload()` method
+     * @test
+     */
+    public function testUpload()
+    {
+        $url = array_merge($this->url, ['action' => 'upload']);
+
+        //GET request
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Banners/upload.ctp');
+
+        //POST request. Tries to upload a banner
+        $this->post(array_merge($url, ['?' => ['position' => 1]]), ['file' => true]);
+        $this->assertResponseOk();
+
+        //Checks the banner has been saved
+        $banner = $this->Banners->find()->last();
+        $this->assertEquals(1, $banner['position_id']);
+        $this->assertEquals('file_to_upload.jpg', $banner['filename']);
+
+        //Deletes all positions, except for the first one
+        $this->Banners->deleteAll(['id >=' => 1]);
+        $this->Banners->Positions->deleteAll(['id >' => 1]);
+
+        //POST request again. Tries to upload a banner
+        //This should also work without the location in the query string, as
+        //  there is only one location
+        $this->post($url, ['file' => true]);
+        $this->assertResponseOk();
+
+        //Checks the banner has been saved
+        $banner = $this->Banners->find()->last();
+        $this->assertEquals(1, $banner['position_id']);
+        $this->assertEquals('file_to_upload.jpg', $banner['filename']);
+    }
+
+    /**
+     * Tests for `download()` method
+     * @test
+     */
+    public function testEdit()
+    {
+        $url = array_merge($this->url, ['action' => 'edit', 1]);
+
+        $this->get($url);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertTemplate(ROOT . 'src/Template/Admin/Banners/edit.ctp');
+
+        $bannerFromView = $this->viewVariable('banner');
+        $this->assertInstanceof('MeCms\Model\Entity\Banner', $bannerFromView);
+        $this->assertNotEmpty($bannerFromView);
+
+        //POST request. Data are valid
+        $this->post($url, ['description' => 'New description for first banner']);
+        $this->assertRedirect(['action' => 'index']);
+        $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
+
+        //POST request. Data are invalid
+        $this->post($url, ['target' => 'invalidTarget']);
+        $this->assertResponseOk();
+        $this->assertResponseNotEmpty();
+        $this->assertResponseContains('The operation has not been performed correctly');
+
+        $bannerFromView = $this->viewVariable('banner');
+        $this->assertInstanceof('MeCms\Model\Entity\Banner', $bannerFromView);
+        $this->assertNotEmpty($bannerFromView);
+    }
+
+    /**
      * Tests for `download()` method
      * @test
      */
     public function testDownload()
     {
-        $url = array_merge($this->url, ['action' => 'download', 1]);
-
-        $this->get($url);
+        $this->get(array_merge($this->url, ['action' => 'download', 1]));
         $this->assertResponseOk();
         $this->assertFileResponse(BANNERS . 'banner1.jpg');
+    }
+
+    /**
+     * Tests for `delete()` method
+     * @test
+     */
+    public function testDelete()
+    {
+        $this->post(array_merge($this->url, ['action' => 'delete', 1]));
+        $this->assertRedirect(['action' => 'index']);
+        $this->assertSession('The operation has been performed correctly', 'Flash.flash.0.message');
     }
 }